requirements.txt
# Core FastMCP Framework
fastmcp
fastapi
uvicorn[standard]
pydantic

# Data Processing
pandas
numpy
openpyxl
pandasql

# Document Parsing
python-docx
python-pptx
pypdf
pdf2image
pytesseract
Pillow

# Semantic Search and ML
sentence-transformers
scikit-learn
chromadb

# Web Scraping and HTTP
requests
beautifulsoup4

# Data Visualization
plotly
pymermaid

# Supabase Client
supabase

# External Connectors (OAuth, encryption)
cryptography>=42.0.0
httpx>=0.27.0

# Environment Variables
python-dotenv

# Evaluation Framework (optional)
# ragas  # For full RAGAS evaluation: pip install ragas datasets
# sacrebleu  # For production BLEU: pip install sacrebleu

# Windows COM Automation (optional, for .ppt files on Windows)
# pywin32